Installed template
I installed the template for landing pages and updated the file structure.
