+++ debtfinder/account/admin.py
@@ -4,6 +4,7 @@ from .models import *
 
 
 # Register your models here.
+admin.site.register(CustomUser)
 admin.site.register(School)
 admin.site.register(Parent)
 admin.site.register(Student)
