+++ debtfinder/account/templates/account/login.html
@@ -5,9 +5,8 @@
 
 <form action="" method="POST">
     {% csrf_token %}
-    <input type="text" name="username" id="username">
+    <input type="email" name="email" id="email">
     <input type="password" name="password" id="password">
-    <input type="submit" value="Submit">
+    <input type="submit" value="Log in">
 </form>
-<p>Don't you have an account? Register <a href="{% url 'parentRegView' %}">here</a></p>
 {% endblock content %}
