+++ debtfinder/templates/ListOfSchools/ListOfSchool.css
@@ -1,0 +1,3 @@
+* {
+background-color:
+}
