+++ debtfinder/templates/index.html
@@ -1,22 +1,18 @@
-<!-- {% extends 'base.html' %}
-
-{% block content %}
-<h1>Landing page</h1> -->
-
 <!DOCTYPE html>
+{% load static %}
 <html lang="en">
 
     <head>
         <meta charset="UTF-8">
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Debtfinder</title>
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
         <link rel="stylesheet" href="style.css">
         <link rel="preconnect" href="https://fonts.googleapis.com">
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
         <link href="https://fonts.googleapis.com/css2?family=Nunito:wght@400;600;700;900&family=Roboto:wght@400;700&display=swap" rel="stylesheet">
-
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+        <link href="{% static 'styles/main.css' %}" rel="stylesheet" type="text/css" media="screen">
+        <title>Debtfinder</title>
     </head>
 
     <body>
@@ -25,7 +21,7 @@
             <nav class="navbar">
 
                 <div class="brand__logo">
-                    <img src="img/LOGO (2).svg" id="logo" alt="brand-logo">
+                    <a href="{% url 'homepage' %}"><img src="{% static 'img/landimg/logo.svg' %}" id="logo" alt="brand-logo"></a>
                 </div>
 
                 <ul class="nav__links">
@@ -51,9 +47,9 @@
 
 
                 <div>
-                    <a href="" id="header-btn-a">Log In
+                    <a href="{% url 'login' %}" id="header-btn-a">Log In
                     </a>
-                    <a href="" id="header-btn-b">Register
+                    <a href="{% url 'createAccount' %}" id="header-btn-b">Register
                     </a>
                 </div>
 
@@ -79,7 +75,7 @@
 
             <div class="image-holder">
 
-                <img src="img/Rectangle 142.svg" id="img" alt="template">
+                <img src="{% static 'img/landimg/hero.svg' %}" id="img" alt="template">
 
             </div>
 
@@ -116,33 +112,33 @@
 
             <div class="logo-container">
                 <div class="top-left">
-                    <img src="img/Rectangle 24.svg" alt="client" id="client">
+                    <img src="{% static 'img/landimg/client1.svg' %}" alt="client" id="client">
                 </div>
 
                 <div class="bottom-left">
-                    <img src="img/Rectangle 27.svg" alt="client" id="client">
+                    <img src="{% static 'img/landimg/client2.svg' %}" alt="client" id="client">
                 </div>
 
 
                 <div class="bottom-right">
-                    <img src="img/Rectangle 30 (1).svg" alt="client" id="client">
+                    <img src="{% static 'img/landimg/client3.svg' %}" alt="client" id="client">
                 </div>
 
                 <div class="bottom-center">
-                    <img src="img/Rectangle 29.svg" alt="client" id="client">
+                    <img src="{% static 'img/landimg/client4.svg' %}" alt="client" id="client">
                 </div>
 
                 <div class="centered">
-                    <img src="img/Rectangle 25.svg" alt="client" id="client">
+                    <img src="{% static 'img/landimg/client5.svg' %}" id="client">
                 </div>
 
                 <div class="center">
-                    <img src="img/Rectangle 26.svg" alt="client" id="client">
+                    <img src="{% static 'img/landimg/client6.svg' %}" id="client">
                 </div>
 
 
                 <div class="top-right">
-                    <img src="img/Rectangle 27.svg" alt="client" id="client">
+                    <img src="{% static 'img/landimg/client7.svg' %}" id="client">
                 </div>
             </div>
 
@@ -180,7 +176,7 @@
 
 
                 <div class="png-image">
-                    <img src="img/Rectangle 85.jpg" id="png" alt="image">
+                    <img src="{% static 'img/landimg/value.jpg' %}" id="png" alt="image">
                 </div>
 
 
@@ -197,17 +193,17 @@
 
             <div class="images">
                 <div class="img-holder">
-                    <img src="img/img1.png" id="imgg">
+                    <img src="{% static 'img/landimg/img1.png' %}" id="imgg">
                     <p>Easy tracking of debts</p>
                 </div>
                 <div class="img-holder">
-                    <img src="img/img2.png" id="imgg">
+                    <img src="{% static 'img/landimg/img2.png' %}" id="imgg">
                     <p>
                         Debtors listing
                     </p>
                 </div>
                 <div class="img-holder">
-                    <img src="img/img3.png" id="imgg">
+                    <img src="{% static 'img/landimg/img3.png' %}" id="imgg">
                     <p>School admin collaboration</p>
                 </div>
             </div>
@@ -222,13 +218,13 @@
             </div>
 
             <div class="previous-arrow">
-                <i><</i></div><divclass="next-arrow"><i>></i>
+                <i></i></div><divclass="next-arrow"><i></i>
             </div>
 
 
             <div class="card-container">
                 <div class="card">
-                    <img src="img/profile pic.png" id="profile-img">
+                    <img src="{% static 'img/landimg/profile1.png' %}" id="profile-img">
                     <h4>Shabnah Saif</h4>
                     <p class="img-p">School Owner</p>
                     <p>A great app for school<span>owners to track debtors. I</span>
@@ -237,7 +233,7 @@
                 </div>
 
                 <div class="card">
-                    <img src="img/profile pic2.png" id="profile-img">
+                    <img src="{% static 'img/landimg/profile2.png' %}" id="profile-img">
                     <h4>Karen Smith</h4>
                     <p class="img-p">School Owner</p>
                     <p>Debt finder have being a
@@ -246,7 +242,7 @@
                 </div>
 
                 <div class="card">
-                    <img src="img/profile pic3.png" id="profile-img">
+                    <img src="{% static 'img/landimg/profile3.png' %}" id="profile-img">
                     <h4>Abubakar Aremu</h4>
                     <p class="img-p">Principal</p>
                     <p>A debtor listing app has<span>
@@ -256,17 +252,13 @@
 
 
                 <div class="card">
-                    <img src="img/profile pic4 (1).png" id="profile-img">
+                    <img src="{% static 'img/landimg/profile4.png' %}" id="profile-img">
                     <h4>Amir  Prakash</h4>
                     <p class="img-p">School Manager</p>
                     <p>I will recommed this tool for<span>
                             any school owner looking to</span>
                         run their school efficiently</p>
                 </div>
-            </div>
-
-            <div class="review-link ">
-                <a href="">See All Review</a>
             </div>
 
         </section>
@@ -364,8 +356,3 @@
     </body>
 
 </html>
-
-
-<!-- 
-{% endblock content %} -->
-
